build: intellisense and validation for env vars
Create a schema which will be used to parse `import.meta.env` and export
the resulting object.

Augment the global `ImportMetaEnv` interface with the inferred type of
the schema.

This took me way too long to figure out because I was getting the
following error in VSCode: "Statements are not allowed in ambient
contexts.ts(1036)". Turns out, the error goes away when you remove the
semicolons within the declaration block.

Only took me a few hours to debug!