refactor(content)!: update "socials" schema
Update the `link` schema to include more properties which can be used to
create anchor elements. Use it to simplify the existing `socials`
schema. Export the data from the "socials" entry from "src/content".

Refactor the `<Contact />` component to create the anchor elements
without the importing a subcomponent and to reflect the new schema
changes.