refactor(layout)!: set `min-height` of `<main>`
Wrap `<slot>` in a `<main>` element which has a `class:list` attribute
which sets the minimum height and the value of the "class" prop.

Remove surrounding `<main>` tags from pages that use `<MainLayout>`.

BREAKING CHANGE: `<MainLayout>` now wraps its `<slot>` with `<main>`
tags; pages that use this component should not include them when passing
child elements to it.