fix: noise background overflowing
On Chromium and Firefox, the background which is `<div>` with a noise
filter inset in the `<main>` element overflows to the footer element,
but without the gradient overlayed on top of it.

To sidestep this problem, the `<body>` element now uses relative
positioning with the `<footer>` positioned at the bottom. Now the
noise background fades down the entire page correctly.