fix: render background pattern in Firefox
Restructure the elements in the `MainLayout` to have the background with
the gradient and pattern to fill the `<body>` element instead of the
`<main>` element.

Place the `<Footer>` back in normal flow since the absolute positioning
was a workaround and is no longer needed.

Use normal mix-blending since since Webkit appears to use different
compositing and blending algorithms which yields different results
compared to Chromium and Firefox.

Slight changes to gradient colors.