ci: add a messy deployment script (#1)
I need to consolidate `commit-subtree.sh` and `deploy.sh` into one
script since the former is really only there for deployment purposes and
to see if my `git-write-subtree` idea would work.

I also need to have the script to be able to parse options given to it
such that I don't need to hard-code values.