refactor: use `<Link>` component for socials (#6)
There is a bug where the icon class specified in the collection entry is
not getting processed by UnoCSS even though the class is added to the
markup. If I don't specify an icon in the collection entry, the `<Link>`
component does not receive that prop and uses the default value which
does get processed by UnoCSS.

This problem must be due to the order of when collection entries get
processed, when an Astro component gets compiled, and when UnoCSS
generates the class.