ci: remove `git push` from script
Remove the `git push` command so we can review the changes that the
script did before we push the deploy branch and trigger a workflow.

The scripts are still a work-in-progress and I am thinking of rewriting
it in TypeScript and run it on Bun. I know that it doesn't make much
sense, but I'm doing it to see if I can do all the building and stuff
locally on my machine and only have GitHub do the actual deployment.

Also it's fun:)