feat: add an `<Icon>` component (#6)
UnoCSS is able to generate classes for Astro components if it can
statically analyze the files (as far as my understanding). So instead of
defining the classes in the collection entry and passing it in as a
prop, we define a dictionary in the component's frontmatter which has
the class names for their respective icons.

This approach was inspired by the `<Icon>` component in the Astro
Starlight package.