refactor: redistribute classes among components
Move the white gradient at the bottom of the page from the `<Perlin>`
component to the `<Footer>` itself and add more padding to it.

Remove classes from the elements inside the `<Perlin>` component like
"mix-blend-normal" and "-z-50".

From the `<MainLayout>` component, pass classes "absolute inset-0 -z-50"
to the `<Perlin>` component. Setting the position to `absolute` creates
a new stacking context for all the child elements in the component,
which means we don't have to specify the z-index for each of them.

Important to note: setting the position to `absolute` resets the
relative position for children just as `relative` does.