refactor: move `<ViewTransitions>` component back
Move the component back because I forgot that it makes every page fades
in and out on Chromium which looks super good, whereas on other browsers
like Safari and Firefox it doesn't make much of a difference.

I'm leaving it in the `<BaseLayout>` component cause it looks awesome on
Chromium.