build(astro): upgrade to astro 3.0.12
Add a "trustedDependencies" property containing `sharp` (the default
image processing service for Astro) because it uses custom install
scripts which makes it incompatible with a normal `bun install`.

For security reasons, Bun **will not** run the lifecycle scripts of
installed dependencies.

The "trustedDependencies" property enables developers to explicitly
define an allowlist of packages that are allowed to run scripts.

Steps for this workaround:
- add the "trustedDependencies" property to the `package.json`
- `rm -rf node_modules bun.lockb`
- `bun install`

See these issues and PRs:
- https://github.com/oven-sh/bun/issues/4549
- https://github.com/oven-sh/bun/issues/2073
- https://github.com/oven-sh/bun/issues/3746