refactor: use `hover` class instead of `@hover`
When running the dev server and when building, UnoCSS keeps reminding me
"The `@hover` variant is experimental and may not follow semver", so
we're just gonna stick with the normal `hover` for now.

I think this also relates to an issue of tapping on the logo on iOS and
iPadOS where the chars adjacent to the char that was tapped still have
sticky hover styles, even though the one that was clicked do not have
them.

So for the sake of consistency, we'll also have stick hover styles for
the tapped char too, until I get around to writing my own CSS styles
that fix the sticky hover problem.