fix(content): use ISO 8601 date-only form
When attempting to run the dev server using the Bun runtime, it throws
an `InvalidContentEntryFrontmatterError` for an entry in the "cv"
collection.

It took me while to find out, but it was because I accidentally provided
a value "March 2022" for the "start" field in the `daterange` schema.
Changing it to "2022-03" solves this issue.

I've no idea why Node does not throw that error.