refactor(photos): use UnoCSS `@screen` directives
On previous version of UnoCSS, there was a warning that constantly
showed up when using the `@unocss/transformer-directives`: "Sourcemap
for <some-file>.astro points to missing source files". Since updating,
the error does not show up anymore when building for production.

Also referencing the responsive breakpoints defined with UnoCSS helps
refactorability.