impr(resume): reverse-chrono one col layout
Decent amount of trying stuff and refactoring went into this...

- use new `<Header>` component and replace the `<Contact>` component
- no more interactive elements and disclosures
- put the skills at the end of the resume
- reorder sections so all the entries that have dates on them are now
  one after the other
- make the root font size 12px when printing.
- other small changes to print styles
- make the width of the actual content 65ch, excluding padding and
  margins

It should be easier now to rapidly make changes and create create custom
tailored versions of resumes when applying to jobs. For example, only
show headings from the "Volunteering" section without the long list of
bullet points to the job can be done in a couple minutes.

Now to automate the PDF saving part...