refactor: throw error if fetch fails
The `ctx.addIssue()` method used in zod refinements and transforms are
used when adding custom validation logic and to customize the resulting
ZodError when attempting to parse invalid input. If a fetch fails due to
some error that is not caused by invalid input or not related to zod at
all then the error thrown should not be an instance of ZodError.