refactor: move render logic into new `src/lib` dir
Planning on using this module not just to generate favicons, but maybe
some OpenGraph images in the future.

Removed caching logic since it is simpler to store generated SVGs in
local vars. We don't need to use JSX in the 'render' module since we can
call `createElement()` from React in the few places we need it.