refactor(css): add `slide-entering-item`
In addition to the `slide-entering-content` class which can be applied
on the parent element, add a `slide-entering-item` class to the list of
selectors.

Refactor markup to use the `slide-entering-item` instead of
`slide-entering-content`.

Leaving both options for flexiblity if I want to extract this stuff out
into a separate file or use in another project.